requirements.txt
Django>=4.2.0
djangorestframework>=3.15.1
python-decouple>=3.8.0
psycopg2-binary>=2.9.6
django-cors-headers>=4.0.0
pillow>=9.10.0
pytest>=7.3.1
pytest-django>=4.5.2
django-environ>=0.4.5
djangorestframework-simplejwt>=5.3.0